requirements.txt
PyYAML
docopt
bottle
